CORE-15586 Split verification from TransactionSignatureService into TransactionSignatureVerificationService
